BISERVER-8103: Removed BlockoutEditor and corresponding validators.  Create a new blockout dialog type
