--- pentaho-gwt-widgets/src/org/pentaho/gwt/widgets/client/controls/schededitor/BlockoutEditor.java
@@ -1,8 +1,0 @@
-package org.pentaho.gwt.widgets.client.controls.schededitor;
-
-import com.google.gwt.user.client.ui.VerticalPanel;
-import org.pentaho.gwt.widgets.client.ui.IChangeHandler;
-
-public class BlockoutEditor //extends AbstractScheduleEditor
-{
-}
